+++ src/core/computedvalue.ts
@@ -108,7 +108,7 @@ export class ComputedValue<T> implements IObservable, IComputedValue<T>, IDeriva
      * This is useful for working with vectors, mouse coordinates etc.
      */
     constructor(options: IComputedValueOptions<T>) {
-        if (!options.get && process.env.NODE_ENV === "production")
+        if (process.env.NODE_ENV === "production" && !options.get)
             return fail("missing option for computed: get")
         this.derivation = options.get!
         this.name = options.name || "ComputedValue@" + getNextId()
